refactor: extract buildDiscoverParams shared util
